refactor: change all IDs to Integers with sequencial auto increment
