+++ backend/src/modules/security/services/TokenService.kt
@@ -5,7 +5,7 @@ import com.auth0.jwt.algorithms.Algorithm
 import com.auth0.jwt.exceptions.JWTCreationException
 import com.auth0.jwt.exceptions.JWTVerificationException
 import com.auth0.jwt.interfaces.DecodedJWT
-import com.davintiproject.backend.modules.security.data.entities.User
+import com.davintiproject.backend.modules.security.domain.entities.User
 import org.springframework.stereotype.Service
 import java.time.Duration
 import java.time.Instant
@@ -17,7 +17,7 @@ class TokenService {
             val algorithm = Algorithm.HMAC256("SUPERSECRET")
 
             return JWT.create()
-                .withSubject(user.id)
+                .withSubject(user.id.toString())
                 .withClaim("name", user.completeName)
                 .withClaim("email", user.email)
                 .withClaim("roles", user.role.name)
